Fix integer overflow in exgcd
The old exgcd can overflow when its arguments are close to the max/min
of the signed integer type. So we convert the arguments to unsigned
integers before computation. This will make sure the value of ta and tb
always decrease in iterations.

This also introduces the sign and unsigned_abs utility functions.