Allow clang-tidy workflow to run on forks
In order to allow clang-tidy-review action to put comments on PRs, we
need to share the github token. This can be achieved by using
pull_request_target event instead of pull_request event. However,
sharing the token with an unknown fork can be dangerous since it grants
them write access. Therefore, the job needs to be approved before it is
executed.